src/requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
asgiref==3.6.0 ; python_version >= '3.7'
black==21.12b0
certifi==2022.12.7 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
django==4.2b1
django-environ==0.10.0
flake8==4.0.1
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0
hpack==4.0.0 ; python_full_version >= '3.6.1'
httpcore==0.16.3 ; python_version >= '3.7'
httpx[http2]==0.23.3 ; python_version >= '3.7'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
idna==3.4
isort==5.12.0
mccabe==0.6.1
mypy-extensions==1.0.0 ; python_version >= '3.5'
pathspec==0.11.0 ; python_version >= '3.7'
platformdirs==3.1.0 ; python_version >= '3.7'
psycopg2-binary==2.9.5
pycodestyle==2.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyflakes==2.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-telegram-bot==20.1
rfc3986[idna2008]==1.5.0
sniffio==1.3.0 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
tomli==1.2.3 ; python_version >= '3.6'
typing-extensions==4.5.0 ; python_version >= '3.10'
